Game Over logic pt. 1, platform, BG music
I added a few death conditions, although, there is more to come. The game is over when the player goes too far left in the screen or falls down the screen. The platforms are now properly rearranged for the rest of the game, and there is background music that fades away when the player dies.